agritech/requirements.txt
asgiref==3.8.1
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.32.0
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
djangorestframework==3.15.2
idna==3.10
isodate==0.7.2
msrest==0.7.1
oauthlib==3.2.2
pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
